[Refactor] svgr 적용 중
