+++ package.json
@@ -22,6 +22,7 @@
     "tailwindcss": "3.3.2"
   },
   "devDependencies": {
+    "@svgr/webpack": "^8.0.1",
     "babel-plugin-styled-components": "^2.1.4",
     "babel-preset-next": "^1.4.0",
     "json-server": "^0.17.3"
