+++ src/assets/img/Background/index.js
@@ -1,2 +1,2 @@
-export { default as Slash } from './back-slash.svg';
-export { default as Wave } from './back-wave.svg';
+// export { default as Slash } from './back-slash.svg?url';
+// export { default as Wave } from './back-wave.svg?url';
