+++ src/assets/img/Icons/index.js
@@ -1,13 +1,13 @@
-export { default as Back } from './ico-back.svg';
+// export { default as Back } from './ico-back.svg?url';
 export { default as FireWorkGreen } from './ico-firework_green.png';
 export { default as FireWorkRed } from './ico-firework_red.png';
 export { default as FireWorkStar } from './ico-firework_star.png';
-export { default as Foot } from './ico-foot.svg';
+// export { default as Foot } from './ico-foot.svg?url';
 export { default as PercentChar } from './ico-percent_char.png';
-export { default as Share } from './ico-share.svg';
-export { default as SnsFaceBook } from './ico-facebook.svg';
-export { default as SnsKakao } from './ico-kakao.svg';
-export { default as SnsTwitter } from './ico-twitter.svg';
-export { default as SnsLink } from './ico-link.svg';
-export { default as TransEn } from './ico-trans_en.svg';
-export { default as TransKo } from './ico-trans_ko.svg';
+// export { default as Share } from './ico-share.svg?url';
+// export { default as SnsFaceBook } from './ico-facebook.svg?url';
+// export { default as SnsKakao } from './ico-kakao.svg?url';
+// export { default as SnsTwitter } from './ico-twitter.svg?url';
+// export { default as SnsLink } from './ico-link.svg?url';
+// export { default as TransEn } from './ico-trans_en.svg?url';
+// export { default as TransKo } from './ico-trans_ko.svg?url';
