+++ src/assets/img/Logo/index.js
@@ -1,2 +1,2 @@
-export { default as LogoBlack } from './logo-black.svg';
-export { default as LogoWhite } from './logo-white.svg';
+export { default as LogoBlack } from './logo-black.svg?url';
+export { default as LogoWhite } from './logo-white.svg?url';
